move `cli` and `utils` to under `h1st`
