+++ h1st/cli/pred_maint.py
@@ -1,7 +1,7 @@
 """H1st Predictive Maintenance / Fault Prediction ("PMFP") CLI."""
 
 
-from ..pred_maint.cli import h1st_pmfp_cli
+from h1st.contrib.pred_maint.cli import h1st_pmfp_cli
 
 
 __all__ = ('h1st_pmfp_cli',)
