+++ h1st/contrib/pred_maint/models/oracle/student/timeseries_dl.py
@@ -21,13 +21,13 @@ from ruamel import yaml
 
 from h1st.model.oracle.student import StudentModeler, Student
 
-from h1st.contrib.utils.data_proc import (PandasFlatteningSubsampler,
-                                          PandasMLPreprocessor,
-                                          ParquetDataset)
-from h1st.contrib.utils.data_proc._abstract import ColsType
-from h1st.contrib.utils.iter import to_iterable
-from h1st.contrib.utils.path import add_cwd_to_py_path
-from h1st.contrib.utils import s3
+from h1st.utils.data_proc import (PandasFlatteningSubsampler,
+                                  PandasMLPreprocessor,
+                                  ParquetDataset)
+from h1st.utils.data_proc._abstract import ColsType
+from h1st.utils.iter import to_iterable
+from h1st.utils.path import add_cwd_to_py_path
+from h1st.utils import s3
 
 from h1st.contrib.pred_maint.data_mgmt import (EquipmentParquetDataSet,
                                                EQUIPMENT_INSTANCE_ID_COL, DATE_COL)   # noqa: E501
