+++ h1st/utils/iter.py
