+++ pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
 name = 'h1st'
-version = '0.1.11'
+version = '0.1.12.dev0'
 description = 'Human-First AI (H1st)'
 authors = ['Aitomatic, Inc. <engineering@aitomatic.com>']
 license = 'Apache-2.0'
@@ -51,7 +51,7 @@ Sphinx-RtD-Theme = '^1.0.0'
 
 
 [tool.poetry.scripts]
-h1st = 'h1st.contrib.cli:h1st_cli'
+h1st = 'h1st.cli:h1st_cli'
 
 
 [build-system]
